Part 20 - Client - Sort - Pagination FUNCIONANDO

+++ controllers/categoryCtrl.js
@@ -1,4 +1,5 @@
 const Category = require('../models/categoryModel')
+const Products = require('../models/productModel')
 
 const categoryCtrl = {
     getCategories: async (req, res) => {
@@ -34,6 +35,9 @@ const categoryCtrl = {
     },
     deleteCategory: async (req, res) => {
         try {
+            const products = await Products.findOne({category: req.params.id})
+            if(products) return res.status(400).json({msg: "Porfavor borre primero todos los products que contengan esta categoría"})
+
             await Category.findByIdAndDelete(req.params.id)
             res.json({msg: "Categoria borrada"})
         } catch (err) {
